+++ src/RLAgent/SharedUtilities.cs
@@ -70,13 +70,13 @@ public class CircularBuffer<T>
 public class MarketData
 {
     public DateTime Timestamp { get; set; }
-    public double Open { get; set; }
-    public double High { get; set; }
-    public double Low { get; set; }
-    public double Close { get; set; }
-    public double Volume { get; set; }
-    public double Bid { get; set; }
-    public double Ask { get; set; }
+    public decimal Open { get; set; }
+    public decimal High { get; set; }
+    public decimal Low { get; set; }
+    public decimal Close { get; set; }
+    public decimal Volume { get; set; }
+    public decimal Bid { get; set; }
+    public decimal Ask { get; set; }
 }
 
 /// <summary>
